requirements.txt
apache-airflow==2.5.0
psycopg2-binary==2.9.3
pandas
sqlalchemy
# openpyxl